requirements-dev.txt
appdirs==1.4.3
asgiref==3.2.5
coverage==5.0.4
distlib==0.3.0
Django==3.0.7
django-cors-headers==3.2.1
django-debug-toolbar==2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tox==3.14.5
virtualenv==20.0.13
whitenoise==5.0.1
